Bug 888799: Display message if callee is busy
Busy calls are now signaled by error, instead of call state. This
patch adds a related message to the TelephonyHelper for informing
the user.

The old code used to play a tone for 3 seconds and close the Call
dialog afterwards. We cannot do this any longer because TonePlayer
expects the Call dialog to be around, which has already been closed
when the error gets handled.

This patch also replaces the related test of the Dialer app by a new
one. The new test verifies that the Dialer app displays the correct
error message when the number is busy.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>