Bug 961986 - [fugu][buri] it took long time to resume music play after ending a phone call
When a call starts, oncall.html opens and the KeypadManager set the priority
of the TonePlayer's audio channel "telephony" so that users can hear tones
when they tap the keypad during a call. We need to reset the channel to
"normal" when hanging up (and thus closing oncall.html) so that other apps can
resume playing audio. Otherwise, we end up waiting for the cycle collector to
kill the audio context, which can take a long time.