Bug 921971 - [DSDS][FDN] Manage FDN list for both SIMs
This fix makes it possible to have independent Fixed Dialing Numbers for each
SIM card in the phone. A new function `getIccByIndex` has been introduced in
`utils.js` that selects the currently selected ICC object without the user
having to specify the particular index, although it allows that too (used in
`sim_dialog.js`, for example.

Instances of `IccHelper` have been replaced by the result of calls to
`getIccByIndex`.